refactor(settings): move Store Tax IDs into the store details block
Promote `TaxIdsField` from `@wcpos/settings` into `@wcpos/ui` so the
free and Pro `StoreDetailsBlock` implementations can both render the
Store Tax IDs section themselves, instead of the parent screen
laying it out separately. Accept a nullable value so consumers can
pass `data?.store_tax_ids` directly without allocating a fresh array
on every render. Export `TaxIdsLabels` so external label objects can
be type-checked against the canonical shape.
