+++ packages/settings/src/screens/general/index.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { isNumber, isString } from 'lodash';
 
 import { PrivacyInfoModal } from '@wcpos/consent';
-import { Button, Callout } from '@wcpos/ui';
+import { type TaxId } from '@wcpos/ui';
 
 import BarcodeSelect from './barcode-select';
 import { StoreDetailsBlock, type StoreDetailsBlockProps } from './store-details-block';
@@ -12,15 +12,9 @@ import UserSelect from './user-select';
 import { FormRow, FormSection } from '../../components/form';
 import Label from '../../components/label';
 import { Skeleton } from '../../components/skeleton';
-import TaxIdsField, {
-	type TaxId,
-	type TaxIdsFieldHandle,
-} from '../../components/tax-ids-field';
 import { Toggle, Checkbox } from '../../components/ui';
 import useSettingsApi from '../../hooks/use-settings-api';
 import { t } from '../../translations';
-
-const STORE_TAX_IDS_DOCS_URL = 'https://wcpos.com/docs/store-tax-ids';
 
 export interface StoreDefaults {
 	store_name: string;
@@ -62,7 +56,6 @@ function getStoreDetailsBlockOverride(): React.ComponentType<StoreDetailsBlockPr
 function General() {
 	const { data, mutate } = useSettingsApi('general');
 	const [privacyInfoOpen, setPrivacyInfoOpen] = React.useState(false);
-	const taxIdsFieldRef = React.useRef<TaxIdsFieldHandle>(null);
 
 	const storeDefaults: StoreDefaults = {
 		store_name: '',
@@ -72,64 +65,12 @@ function General() {
 		...(data?.store_defaults ?? {}),
 	};
 
-	const storeTaxIdsDescription = (
-		<>
-			{t('settings.store_tax_ids_section_description')}{' '}
-			<a
-				href={STORE_TAX_IDS_DOCS_URL}
-				target="_blank"
-				rel="noreferrer noopener"
-				className="wcpos:text-wp-admin-theme-color wcpos:underline"
-			>
-				{t('settings.store_tax_ids_learn_more')}
-			</a>
-		</>
-	);
-
 	const StoreDetailsBlockOverride = getStoreDetailsBlockOverride();
 	const ResolvedStoreDetailsBlock = StoreDetailsBlockOverride ?? StoreDetailsBlock;
 
 	return (
 		<>
 			<ResolvedStoreDetailsBlock data={data} mutate={mutate} storeDefaults={storeDefaults} />
-			<FormSection
-				title={t('settings.store_tax_ids_section_title')}
-				description={storeTaxIdsDescription}
-				headerRight={
-					<Button variant="outline" onClick={() => taxIdsFieldRef.current?.addRow()}>
-						{t('settings.store_tax_ids_add')}
-					</Button>
-				}
-				divider
-			>
-				<Callout
-					status="info"
-					title={t('settings.store_tax_ids_callout_title')}
-					className="wcpos:mb-4"
-				>
-					{t('settings.store_tax_ids_tip')}
-				</Callout>
-				<TaxIdsField
-					ref={taxIdsFieldRef}
-					value={Array.isArray(data?.store_tax_ids) ? data.store_tax_ids : []}
-					onChange={(store_tax_ids) => {
-						mutate({ store_tax_ids });
-					}}
-					labels={{
-						add: t('settings.store_tax_ids_add'),
-						type: t('settings.store_tax_ids_type'),
-						value: t('settings.store_tax_ids_value'),
-						country: t('settings.store_tax_ids_country'),
-						countryPlaceholder: t('settings.store_tax_ids_country_placeholder'),
-						countrySearchPlaceholder: t('settings.store_tax_ids_country_search'),
-						countryNoResults: t('settings.store_tax_ids_country_no_results'),
-						countryClear: t('settings.store_tax_ids_country_none'),
-						label: t('settings.store_tax_ids_label'),
-						remove: t('settings.store_tax_ids_remove'),
-						empty: t('settings.store_tax_ids_empty'),
-					}}
-				/>
-			</FormSection>
 			<FormSection title={t('settings.products_section_title')} divider>
 				<FormRow>
 					<Label tip={t('settings.pos_only_products_tip')}>
