+++ packages/settings/src/screens/general/store-details-block.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 
 import { isString } from 'lodash';
 
-import { Callout } from '@wcpos/ui';
+import { Button, Callout, TaxIdsField, type TaxId, type TaxIdsFieldHandle } from '@wcpos/ui';
 
 import { FormRow, FormSection } from '../../components/form';
 import { TextInput, TextArea } from '../../components/ui';
@@ -13,6 +13,7 @@ import type { StoreDefaults } from './index';
 
 const UPGRADE_URL = 'https://wcpos.com/pro';
 const UPGRADE_PLACEMENT = 'general_store_details';
+const STORE_TAX_IDS_DOCS_URL = 'https://wcpos.com/docs/store-tax-ids';
 
 export interface StoreDetailsBlockProps {
 	data?: {
@@ -20,6 +21,7 @@ export interface StoreDetailsBlockProps {
 		store_phone?: unknown;
 		store_email?: unknown;
 		policies_and_conditions?: unknown;
+		store_tax_ids?: unknown;
 	};
 	mutate: (data: Record<string, unknown>) => void;
 	storeDefaults: StoreDefaults;
@@ -30,9 +32,25 @@ export interface StoreDetailsBlockProps {
  * `general.store_details_block` registry slot when stores exist.
  */
 export function StoreDetailsBlock({ data, mutate, storeDefaults }: StoreDetailsBlockProps) {
+	const taxIdsFieldRef = React.useRef<TaxIdsFieldHandle>(null);
+
 	React.useEffect(() => {
 		captureUpgradeCtaViewed(UPGRADE_PLACEMENT);
 	}, []);
+
+	const storeTaxIdsDescription = (
+		<>
+			{t('settings.store_tax_ids_section_description')}{' '}
+			<a
+				href={STORE_TAX_IDS_DOCS_URL}
+				target="_blank"
+				rel="noreferrer noopener"
+				className="wcpos:text-wp-admin-theme-color wcpos:underline"
+			>
+				{t('settings.store_tax_ids_learn_more')}
+			</a>
+		</>
+	);
 
 	return (
 		<>
@@ -41,6 +59,25 @@ export function StoreDetailsBlock({ data, mutate, storeDefaults }: StoreDetailsB
 				description={t('settings.store_details_section_description')}
 				divider
 			>
+				<Callout
+					status="info"
+					title={t('settings.store_details_upgrade_callout_title')}
+					className="wcpos:mt-4"
+				>
+					<p className="wcpos:m-0">
+						{t('settings.store_details_upgrade_to_pro')}
+					</p>
+					<p className="wcpos:mt-2 wcpos:mb-0">
+						<a
+							href={UPGRADE_URL}
+							target="_blank"
+							rel="noreferrer noopener"
+							onClick={() => captureUpgradeCtaClicked(UPGRADE_PLACEMENT, UPGRADE_URL)}
+						>
+							{t('common.upgrade_to_pro')}
+						</a>
+					</p>
+				</Callout>
 				<FormRow label={t('settings.store_name')}>
 					<TextInput
 						value={isString(data?.store_name) ? data.store_name : ''}
@@ -76,25 +113,44 @@ export function StoreDetailsBlock({ data, mutate, storeDefaults }: StoreDetailsB
 						onChange={(event) => mutate({ policies_and_conditions: event.target.value })}
 					/>
 				</FormRow>
+			</FormSection>
+			<FormSection
+				title={t('settings.store_tax_ids_section_title')}
+				description={storeTaxIdsDescription}
+				headerRight={
+					<Button variant="outline" onClick={() => taxIdsFieldRef.current?.addRow()}>
+						{t('settings.store_tax_ids_add')}
+					</Button>
+				}
+				divider
+			>
 				<Callout
 					status="info"
-					title={t('settings.store_details_upgrade_callout_title')}
-					className="wcpos:mt-4"
+					title={t('settings.store_tax_ids_callout_title')}
+					className="wcpos:mb-4"
 				>
-					<p className="wcpos:m-0">
-						{t('settings.store_details_upgrade_to_pro')}
-					</p>
-					<p className="wcpos:mt-2 wcpos:mb-0">
-						<a
-							href={UPGRADE_URL}
-							target="_blank"
-							rel="noreferrer noopener"
-							onClick={() => captureUpgradeCtaClicked(UPGRADE_PLACEMENT, UPGRADE_URL)}
-						>
-							{t('common.upgrade_to_pro')}
-						</a>
-					</p>
+					{t('settings.store_tax_ids_tip')}
 				</Callout>
+				<TaxIdsField
+					ref={taxIdsFieldRef}
+					value={Array.isArray(data?.store_tax_ids) ? (data.store_tax_ids as TaxId[]) : null}
+					onChange={(store_tax_ids) => {
+						mutate({ store_tax_ids });
+					}}
+					labels={{
+						add: t('settings.store_tax_ids_add'),
+						type: t('settings.store_tax_ids_type'),
+						value: t('settings.store_tax_ids_value'),
+						country: t('settings.store_tax_ids_country'),
+						countryPlaceholder: t('settings.store_tax_ids_country_placeholder'),
+						countrySearchPlaceholder: t('settings.store_tax_ids_country_search'),
+						countryNoResults: t('settings.store_tax_ids_country_no_results'),
+						countryClear: t('settings.store_tax_ids_country_none'),
+						label: t('settings.store_tax_ids_label'),
+						remove: t('settings.store_tax_ids_remove'),
+						empty: t('settings.store_tax_ids_empty'),
+					}}
+				/>
 			</FormSection>
 		</>
 	);
